Added Iron Strike
Added the common attack Iron Strike. This card costs 2 and does 12 damage plus a Dexterity bonus. Additionally, I had to debug the Steamroll card. Both Iron Strike and Steamroll were not receiving a weakness debuff because their damage is not tied to the variable "damage". Now, all weakness calculations are done in the class.